Move StatusModifier into own module
Splitting things into modules is always nice.  Also its much easier to
reexport it for the users of the crate.
