+++ src/lib.rs
@@ -12,7 +12,7 @@ extern crate spaceapi;
 
 pub use spaceapi as api;
 pub use server::SpaceapiServer;
-pub use server::handlers::{StatusModifier, StateFromPeopleNowPresent};
+pub use server::modifiers;
 
 mod server;
 pub mod datastore;
