+++ src/server/handlers.rs
@@ -1,7 +1,6 @@
 //! Handlers for the server.
 
 use std::collections::BTreeMap;
-use std::any::Any;
 
 use rustc_serialize::json::{Json, ToJson};
 use iron::prelude::*;
@@ -15,6 +14,7 @@ use ::api;
 use ::api::optional::Optional;
 use ::datastore;
 use ::sensors;
+use ::modifiers;
 
 
 #[derive(Debug)]
@@ -32,38 +32,11 @@ impl ToJson for ErrorResponse {
     }
 }
 
-
-/// `StatusModifier`s are used to modify the status
-pub trait StatusModifier: Send + Sync + Any {
-    /// Called after all registered sensors are read
-    fn modify(&self, status: &mut api::Status);
-}
-
-pub struct StateFromPeopleNowPresent;
-
-impl StatusModifier for StateFromPeopleNowPresent {
-    fn modify(&self, status: &mut api::Status) {
-        // Update state depending on number of people present
-        let people_now_present: Option<u64> = status.sensors.as_ref()
-            .and_then(|sensors| sensors.people_now_present.as_ref())
-            .map(|people_now_present| people_now_present[0].value)
-            .into();
-        if let Some(count) = people_now_present {
-            status.state.open = Some(count > 0);
-            if count == 1 {
-                status.state.message = Optional::Value(format!("{} person here right now", count));
-            } else if count > 1 {
-                status.state.message = Optional::Value(format!("{} people here right now", count));
-            }
-        }
-    }
-}
-
 pub struct ReadHandler {
     status: api::Status,
     datastore: datastore::SafeDataStore,
     sensor_specs: sensors::SafeSensorSpecs,
-    status_modifiers: Vec<Box<StatusModifier>>,
+    status_modifiers: Vec<Box<modifiers::StatusModifier>>,
 }
 
 impl ReadHandler {
